refactor: geth tracer layout
